Re-enable commands for local repos.
I've changed the commands to `newrepo` and `deleterepo`, as I think
it is less confusing to users, who may expect it to be mean "new project"
or "delete project", both of which would be reasonable commands to have.

I also removed the dependency on FileSystem.absPath.  Easier to use the straight
`Path` API as we can't guarantee which Haxe version is being used.