Prevent external diff tool launch.
If you have an graphical diff tool defined in global or repo level
gitconfig this command will launch it rather then the text based
expected by the tool.

Fix is to add the switch —no-ext-diff fixes this, around since git 1.7.
See this for example of why this is an issue:
    https://github.com/pre-commit/pre-commit/issues/545