Remember capitalisation of library when installed
Addresses #529

This way, list shows the correct version of the name, and path falls
back to the correct version when no haxelib.json is found.

The "correct" version is the way the name in haxelib.json is
capitalized, or if no such file exists it is the way the parameter to
the `dev`/`git`/`hg` command was capitalized. Additionally, if
the parameter is an alias, then it is used instead as well.

It is updated everytime a version is set as current. However, officially
released library version names are preferred over the capitalization of
`dev`, `git`, or `hg` versions.

For example, after:
`haxelib dev LIBRARY <path>`
`haxelib install library 1.0.0`

If `library` is the version in `haxelib.json`, then that is preferred
over `LIBRARY` after we install from the server.

Throw error in `path` command when two versions of the same lib are
queried even if the two versions are capitalized differently.
- e.g. `LiBrArY:1.2.0` and `library:1.3.0`

Ensure that haxelib run sets HAXELIB_RUN_NAME to name field in
`haxelib.json`.

Ignore capitalized directories in `list`.