[tests] Update vcs unit tests
Updating should not do anything if only a commit/tag was specified at
install.

Added a new test for cloning only with a commit hash (and later
updating).

Merged the two tests for update with local changes, as the logic for
confirming local change reset is now performed externally which leads to
cleaner tests here.