use transaction in processSubmit()
Hopefully it is now safe for most cases. The submission process is problematic - the client uses a pretty short timeout before retrying remoting calls, thus there could be multiple calls of processSubmit() run in parallel if it took a bit too long.

The use of transaction in this commit is far from perfect. Errors can happen in many steps, we should catch them and call rollback(). Or maybe auto abort them in the db. The client may still believe the submission failed and/or created multiple entries of the same version in the db.