Fix subdir checks for git/hg installs
It may be desired to omit the subdir argument but pass in a tag, but
currently this is not possible and it is necessary to pass an empty
string as the subdir. This should not result in a dev path being set.

There is an example that requires this behaviour in the newly added
integration tests, see tests.integration.TestVcs.testVcsUpdateTag().