Revert back to using DBCache (#663)
This fixes the (or at least many) memory usage issues of the haxelib server. The cache is meant to store the results of unzipping the library zip archives. Without the cache, the server unzips a given library's .zip on every single request that requires any information about the library, including the overview page of a given library ( e.g. https://lib.haxe.org/p/hxcpp/).