Remove site dependency in hxml target flag test
The `testInstallHxmlWithBackend` test require lib.haxe.org to be up so
it can fetch the server's hxcpp version, which is not good if we are
redeploying the server because it's down.

The new test just tests the LibFlagData.fromHxml function directly, to
ensure no duplication there. We could also add an integration test in
future.