Disable updates with git/hg install without branch
If only a commit or tag was specified, there is no simple way to
determine what an update is meant to do. The commmit may belong to
multiple branches, and there may be multiple newer tags which have
different purposes (e.g. stable vs alpha release tags). It makes more
sense to simply lock to the specified commit when a user installs this
way.

Also, the git show-branch method would not work with a shallow git
history.

For hg, to achieve the same behaviour as git it is necessary to strip
any newer changesets to the checked out one and also remove the
reference to the upstream repository so that no further updates are
pulled. This is not ideal as it breaks Hg.getOriginUrl, but that should
be saved already in .hgdata so shouldn't be called.