Fix -rc version bug for repo api as well
`-rc` versions were still being reported as `-preview` for repo api
calls. The server flag is not defined for the repo api, only for other
parts of the server. Adding `haxelib_api` here solves the issue.

Also, the Preview enum is mapped directly to the database field. This
means to work properly, any new fields have to be added after all
existing fields.