Add support for eval build of haxelib in makefile
This installs haxelib as an sh script that runs haxelib in eval mode,
with the haxelib code cached in an hxb archive.

The makefile can now switch between the neko install and the eval
install using the HAXELIB_TARGET variable. It caches the last used value
in a file and invalidates if the next build changes the target.